refactor: admin dashboard to server components with auth wrapper
- Move from client-side data fetching to server components with proper
authentication.
- Improve notification handling across components and
enhance component separation for better maintainability.
